Feat: FCM TOPIC & DEVICE TOKEN
Automatically subscribing device and listening to the topic "all". This will be the broadcast.

Capturing device token so that we can be able to pass it on to the api for saving.

The device token will have to be uniquly saved in the db so thet it can be used to target specific user.